[Del] Delete origin and replace refac to origin
